+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Add `paginate` local directive ([#17](https://github.com/marp-team/marpit/pull/17))
 * Make Unicode emoji printable ([#18](https://github.com/marp-team/marpit/pull/18))
+* Prevent style injections ([#19](https://github.com/marp-team/marpit/pull/19))
 
 ## v0.0.4 - 2018-05-05